bug fixed for jmptbl.

+++ src/backend/vm/vmrunlir.c
@@ -827,7 +827,7 @@ static int run_vm_by_lir(struct vm_program *prog, int64_t ip, uint8_t *stack, in
         NEXT();
     }
     VM_CASE_(VM_JMPTBL): {
-        uint64_t index;
+        uint32_t index;
         POPI(index, uint32_t, /* no mask */);
         ip += index + 1;
         ip += base[ip]->d.addr.index;
